backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
Flask-Mail==0.9.1
marshmallow==3.19.0
Werkzeug==2.3.6
python-dotenv==1.0.0
